added unsign_itoa_base.c and length.c

+++ ft_getsize.c
@@ -12,14 +12,14 @@
 
 #include "libft.h"
 
-int		ft_getsize(int value, int base)
+int		ft_getsize(long long value, int base)
 {
 	int i;
 
 	i = 0;
 	if (value < 0)
 		value = value * (-1);
-	while (value >= base)
+	while (value >= (long long)base)
 	{
 		i++;
 		value = value / base;
